Hash email, origin and some randomness to get a session key
This removes the email and origin from the link in the email loop. We can also
use this as the "state" parameter for authentication against famous IdPs. By
using this as the key for our state in Redis, so that different login attempts
for the same email/origin can be distinguished even in the short term.