Move creating error response to lib.rs.
This allows us some more flexibility in how to respond and log errors.

Because we now consume `BrokerError::Input` and turn it into an `Ok`, we
no longer log these as errors.

In addition, because we moved the `$body` expansion in `broker_handler!`
to its own method, the compiler is now picking up on some places where
`$req` was unused.

Eventually, we can use request context to report errors back to the RP.