Validate `response_type` and `response_mode`.
This is a small breaking change, but it improves our compliance with
OAuth2 and OIDC.

The `response_type` parameter is always required.

The default `response_mode` is `fragment`, as per OIDC Core 3.2.2.5, so
`form_post` needs to be explicitely specified in our case. This adds a
default to our parsing, to follow the spec, and so we can perhaps
implement it later.