Allow provider details to be set incrementally
This will allow for setting public details, like the issuer domain,
discovery URL, and client ID in a config file, while keeping the client
secret in an environment variable.

Using a builder also allows us to end up with a concrete Provider struct
at the end of configuration processing, which is easier to handle than a
struct with Option members.