Implement implicit flow for Google.
Google doesn't support `response_mode=form_post`, so we add a small page
inbetween to capture the fragment parameters and POST them to us.

This removes a roundtrip on the server side, but adds a roundtrip on the
client side. However, this also eliminates lingering access tokens, and
maybe even preps us for supporting more OIDC providers in the future.