Idea for integrating a 'Deploy to Heroku' button
The resulting free Heroku dyno has an associated Redis instance and the ability
to send emails via SendGrid's SMTP interface. Email addons in the Heroku
marketplace don't seem to be consistent with respect to the env vars they
populate. I worked around this by specifically checking for SendGrid
credentials, though a more versatile solution capable of also handling Mailgun,
Postmark, etc., would be preferable (similar to the way Redis is handled now).

This is a proof of concept so I didn't try to include all of the possible
portier-broker env vars in app.json. The minimally necessary values for getting
the server up & running are present as well as a few extras for making things
nicer.

Related to #88