Serialization and session refactor
This commit modifies several places to do serialization using types instead of
plucking apart a JSON Value. It touches on HTTP IdP interactions and session
handling.

The JWT stuff is left as is, because it may in the future use an external
library instead.

In addition, this replaces the private scheme for Google with a separate rel.